+++ lib/FilDirection.dart
@@ -1,5 +1,8 @@
-
+/// Defines the direction in which the circular progress is filled.
 enum FillDirection {
+  /// Progress is filled in a clockwise direction.
   clockwise,
+
+  /// Progress is filled in a counter-clockwise direction.
   counterClockwise,
 }