+++ CHANGELOG.md
@@ -1,3 +1,24 @@
-## 0.0.1
+## 1.3.0
+* add font family parameter
+## 1.2.0
+
+* update flutter 
+* update gradle
+* update kotlin
+* maka round progress indicator 
+
+## 1.0.4
+
+* Initial development release
+
+## 1.0.3
+
+* Initial development release
+
+## 1.0.2
+
+* Initial development release
+
+## 1.0.1
 
 * TODO: Describe initial release.